fix(homestay): format homestay response

+++ huta-domain/src/main/java/nnt/com/domain/aggregates/model/dto/response/HomestayResponse.java
@@ -42,7 +42,7 @@ public class HomestayResponse {
     boolean isAds;
     Map<String, Double> map;
     //    List<Room> rooms;
-    double rating;
+    float rating;
     String typeHomestay;
     Integer version;
